src/site/lvkun/balanced_binary_tree/SolutionTest.java
package site.lvkun.balanced_binary_tree;

import org.junit.Assert;
import org.junit.Test;
import site.lvkun.common.TreeNode;

public class SolutionTest {
    @Test
    public void testIsBalanced() {
        Solution solution = new Solution();

        TreeNode root;

        root = new TreeNode(3);
        root.left = new TreeNode(9);
        root.right = new TreeNode(20);
        root.right.left = new TreeNode(15);
        root.right.right = new TreeNode(7);

        Assert.assertTrue(solution.isBalanced(root));

        root = new TreeNode(1);
        root.left = new TreeNode(2);
        root.left.left = new TreeNode(3);
        root.left.right = new TreeNode(3);
        root.left.left.left = new TreeNode(4);
        root.left.left.right = new TreeNode(4);
        root.right = new TreeNode(2);

        Assert.assertFalse(solution.isBalanced(root));
    }
}
